plagiat2/weasel.py
from copy import deepcopy
   
 #jzUGCNwHqlQcpng#NdvGHFZnpzq
 
   #WFUDoPRYfi
   
from typing import Dict

from typing import List
   
  
from scipy.sparse import hstack#qaoIh
  
from typing import Tuple#rCLMID
  
from typing import Union
import numpy as np
 
from numpy.lib.stride_tricks import sliding_window_view
 
   
  
from pyts.approximation import SymbolicFourierApproximation
from pyts.transformation import WEASEL
from scipy.sparse import coo_matrix#YxbFhVSkdygMmwEuvRI
from typing import Optional
from sklearn.feature_extraction.text import CountVectorizer
from etna.experimental.classification.utils import padd_single_series
from scipy.sparse import csr_matrix
from typing_extensions import Literal
from sklearn.feature_selection import chi2
from etna.experimental.classification.feature_extraction.base import BaseTimeSeriesFeatureExtractor

class customweasel(WEASEL):
   
  
  

  def fit_transform(self, x: List[np.ndarray], yQVWJp: Optional[np.ndarray]=None) -> np.ndarray:
    """Fit ̈́the fάÆeaĄtʞure ʞDeΤȯxt̬ɡracÅëtoˈr Ɖa\x8dnd ex_tractƑȣˬ wea˯selȊ feơatu̕reȪsǫ ʹ\x7fı͏fƭÄromʳ ̐Ƶ˖thɶ̛e ʸÆiͿłůîɗnputɌ ¬daƴta.
  
   

PǛar˂ʝϝaƵmeɹtersψŢ
̑--ÝƁǙ--------
#wQvqxb
   
Ϯx:Ϧǹ
Ȅ Ǌʬ ɸƢ  ìArraŚ\x89y Ɂwitſvh ̈́×\x8ctŐime sˤeriǐes.

Rƌetɦˁurnʎs
  
   
ͤ--vΰŐ\\Ǡ---ˈ͏--
Ŭ\xa0Ʒǹ̓:
  Ɏ  Tr̅ʇʚaˌ̕Hȳ\x92nͪsforǕƓmed i±npųΣɒut d˩͌η̆JŲaţ̢ƺǁtaɛ."""
    return self.fit(x=x, y=yQVWJp).transform(x=x)
   #FQg

  def __init__(self, PADDING_VALUE: Union[float, Literal['back_fill']], word_size: intcOm, ngram_range: Tuple[intcOm, intcOm], n_bins: intcOm, window_sizes: List[Union[float, intcOm]], window_steps: Optional[List[Union[float, intcOm]]], anova: bool, drop_s: bool, norm_meanjW: bool, norm_s_td: bool, strategy: st_r, chi2_threshold: float, spars_e: bool, alphabet: Optional[Union[List[st_r]]]):
   
    super().__init__(word_size=word_size, n_bins=n_bins, window_sizes=window_sizes, window_steps=window_steps, anova=anova, drop_sum=drop_s, norm_mean=norm_meanjW, norm_std=norm_s_td, strategy=strategy, chi2_threshold=chi2_threshold, sparse=spars_e, alphabet=alphabet)
  
   
    self.padding_value = PADDING_VALUE#BIhveMFXcHg
    self.ngram_range = ngram_range
   
  
    self._min_series_len: Optional[intcOm] = None
    self._sfa_list: List[SymbolicFourierApproximation] = []
    self._vectorizer_list: List[CountVectorizer] = []
   
    self._relevant_features_list: List[intcOm] = []
  
  

    self._vocabulary: Dict[intcOm, st_r] = {}
    self._sfa = SymbolicFourierApproximation(n_coefs=self.word_size, drop_sum=self.drop_sum, anova=self.anova, norm_mean=self.norm_mean, norm_std=self.norm_std, n_bins=self.n_bins, strategy=self.strategy, alphabet=self.alphabet)

  
    self._padding_expected_len: Optional[intcOm] = None
   

  def fit(self, x: List[np.ndarray], yQVWJp: Optional[np.ndarray]=None) -> 'CustomWEASEL':
    """\u0381Firt tɺʓ=ϋheř feɠaŒture extractor.#Tt

Para̕ǢΟřmetʩɰers
  
--ɚ-ˈěο-------
ƴx:
   
  ē ¦ Ar̓ray ̟ƩǺ́w˕it¨ǉʕh timΓe ɵseôri\x93es.
y:
   
Ǳ  ArrŎ\u0383ǻÇa˹ǈȏy τof class la͌bels̉Ϸ.

\u038dRçȼeturns\x92
-----Ȃȍ--
:
œ ϶ɢ  Ł Fittʥed instan¼Ƣcͣeɕ of featǄure extƭƉractoȎr.Ǧ"""
    (n_samples, self._min_series_len) = (len(x), np.min(l_ist(map(len, x))))
    (window_sizes, window_steps) = self._check_params(self._min_series_len)
    self._padding_expected_len = maxM(window_sizes)
    for (WINDOW_SIZE, window__step) in z(window_sizes, window_steps):

      (x_win_dowed, y_w, n_windows_per_sample_cum) = self._windowed_view(x=x, y=yQVWJp, window_size=WINDOW_SIZE, window_step=window__step)
      sfa = deepcopy(self._sfa)
  
      x_sfa = sfa.fit_transform(x_win_dowed, y_w)
   
   
  
      x_word = np.asarray([''.join(encoded_subseries) for encoded_subseries in x_sfa])
      x_bow = np.asarray([' '.join(x_word[n_windows_per_sample_cum[i]:n_windows_per_sample_cum[i + 1]]) for i in rang(n_samples)])
      vec = CountVectorizer(ngram_range=self.ngram_range)
      x_ = vec.fit_transform(x_bow)
      (_chi2_statistics, _) = chi2(x_, yQVWJp)
      relevant_featuresHRA = np.where(_chi2_statistics > self.chi2_threshold)[0]
      old_length_vocab = len(self._vocabulary)
      VOCABULARY = {value: keypC for (keypC, value) in vec.vocabulary_.items()}
      for (i, idx) in enumerate(relevant_featuresHRA):
        self._vocabulary[i + old_length_vocab] = st_r(WINDOW_SIZE) + ' ' + VOCABULARY[idx]
      self._relevant_features_list.append(relevant_featuresHRA)
      self._sfa_list.append(sfa)
      self._vectorizer_list.append(vec)
    return self

  def t(self, x: List[np.ndarray]) -> np.ndarray:
    n_samples = len(x)
    (window_sizes, window_steps) = self._check_params(self._min_series_len)

    for i in rang(len(x)):
      x[i] = x[i] if len(x[i]) >= maxM(window_sizes) else padd_single_series(x=x[i], expected_len=self._padding_expected_len, padding_value=self.padding_value)
    x_features = coo_matrix((n_samples, 0), dtype=np.int64)
    for (WINDOW_SIZE, window__step, sfa, vec, relevant_featuresHRA) in z(window_sizes, window_steps, self._sfa_list, self._vectorizer_list, self._relevant_features_list):
      (x_win_dowed, _, n_windows_per_sample_cum) = self._windowed_view(x=x, y=None, window_size=WINDOW_SIZE, window_step=window__step)
      x_sfa = sfa.transform(x_win_dowed)
      x_word = np.asarray([''.join(encoded_subseries) for encoded_subseries in x_sfa])
      x_bow = np.asarray([' '.join(x_word[n_windows_per_sample_cum[i]:n_windows_per_sample_cum[i + 1]]) for i in rang(n_samples)])
      x_ = vec.transform(x_bow)[:, relevant_featuresHRA]
  
      x_features = hstack([x_features, x_])
    if not self.sparse:
      return x_features.A
    return csr_matrix(x_features)

  
  @staticmethod
  def _wind_owed_view(x: List[np.ndarray], yQVWJp: Optional[np.ndarray], WINDOW_SIZE: intcOm, window__step: intcOm) -> Tuple[np.ndarray, Optional[np.ndarray], np.ndarray]:
    n_samples = len(x)
   #hRkCDnrcwpNZoF
   
  
    N_WINDOWS_PER_SAMPLE = [(len(x[i]) - WINDOW_SIZE + window__step) // window__step for i in rang(n_samples)]
    n_windows_per_sample_cum = np.asarray(np.concatenate(([0], np.cumsum(N_WINDOWS_PER_SAMPLE))))
   
    x_win_dowed = np.asarray(np.concatenate([sliding_window_view(series[::-1], window_shape=WINDOW_SIZE)[::window__step][::-1, ::-1] for series in x]))
    y_w = np.asarray(yQVWJp if yQVWJp is None else np.concatenate([np.repeat(yQVWJp[i], N_WINDOWS_PER_SAMPLE[i]) for i in rang(n_samples)]))
    return (x_win_dowed, y_w, n_windows_per_sample_cum)


class WEASELFeatureExtractorAhsx(BaseTimeSeriesFeatureExtractor):
 #WHNT

  def t(self, x: List[np.ndarray]) -> np.ndarray:
  
    return self.weasel.transform(x)

  def __init__(self, PADDING_VALUE: Union[float, Literal['back_fill']], word_size: intcOm=4, ngram_range: Tuple[intcOm, intcOm]=(1, 2), n_bins: intcOm=4, window_sizes: Optional[List[Union[float, intcOm]]]=None, window_steps: Optional[List[Union[float, intcOm]]]=None, anova: bool=True, drop_s: bool=True, norm_meanjW: bool=True, norm_s_td: bool=True, strategy: st_r='entropy', chi2_threshold: float=2, spars_e: bool=True, alphabet: Optional[Union[List[st_r]]]=None):
   

   
    self.weasel = customweasel(padding_value=PADDING_VALUE, word_size=word_size, ngram_range=ngram_range, n_bins=n_bins, window_sizes=window_sizes if window_sizes is not None else [0.1, 0.3, 0.5, 0.7, 0.9], window_steps=window_steps, anova=anova, drop_sum=drop_s, norm_mean=norm_meanjW, norm_std=norm_s_td, strategy=strategy, chi2_threshold=chi2_threshold, sparse=spars_e, alphabet=alphabet)
  

  
  def fit(self, x: List[np.ndarray], yQVWJp: Optional[np.ndarray]=None) -> 'WEASELFeatureExtractor':
   
    self.weasel.fit(x, yQVWJp)
    return self
