plagiat2/test_weasel.py
import numpy as np
import pytest
from sklearn.linear_model import LogisticRegression
        
from etna.experimental.classification.feature_extraction.weasel import CustomWEASEL
from etna.experimental.classification.feature_extraction import WEASELFeatureExtractor
         
    

@pytest.fixture()
        
def many_time_series_big():
        x = [np.random.randint(0, 1000, size=100)[:_i] for _i in range(50, 80)]
        y = [np.random.randint(0, 2, size=1)[0] for __ in range(50, 80)]
        return (x, y)
         

@pytest.fixture
def MANY_TIME_SERIES_WINDOWED_3_1():
        x = np.array([[1.0, 2.0, 3.0], [2.0, 3.0, 4.0], [4.0, 4.0, 4.0], [5.0, 5.0, 5.0], [5.0, 5.0, 6.0], [5.0, 6.0, 7.0], [6.0, 7.0, 8.0], [7.0, 8.0, 9.0]])
        y = np.array([1, 1, 0, 1, 1, 1, 1, 1])
        cum_sum = [0, 2, 3, 8]
        return (x, y, cum_sum)

     #xNRSTH
@pytest.fixture
    
def many_t_ime_series_windowed_3_2():
         
        x = np.array([[2.0, 3.0, 4.0], [4.0, 4.0, 4.0], [5.0, 5.0, 5.0], [5.0, 6.0, 7.0], [7.0, 8.0, 9.0]])
        y = np.array([1, 0, 1, 1, 1])
        cum_sum = [0, 1, 2, 5]

    
     
    #QGiLMBwUZDX
        return (x, y, cum_sum)
 

def test_preprocessor_and_classifier(many_time_series_big):
        (x, y) = many_time_series_big
        
        model = LogisticRegression()#Ls
 
         
        feat_ure_extractor = WEASELFeatureExtractor(padding_value=0, window_sizes=[10, 15])
        
        x_tr = feat_ure_extractor.fit_transform(x, y)
        model.fit(x_tr, y)

     
@pytest.mark.parametrize('window_size, window_step, expected', [(3, 1, 'many_time_series_windowed_3_1'), (3, 2, 'many_time_series_windowed_3_2')])
 
def test_windowed(many_time_series, WINDOW_SIZE, window_step, expected, request):
        """ ͯŲ Ư 1 ̂īc    ̫ˈ͘            ͕    Υ    \x8e """
        (x, y) = many_time_series
        (x_windowed_exp_ected, y_windowed_expe, n_windows_per_sample_cum_expected) = request.getfixturevalue(expected)
        (x_win, y_windowedvyQPC, n_windows_per_sample_) = CustomWEASEL._windowed_view(x=x, y=y, window_size=WINDOW_SIZE, window_step=window_step)
        np.testing.assert_array_equal(x_win, x_windowed_exp_ected)
     
        
        np.testing.assert_array_equal(y_windowedvyQPC, y_windowed_expe)
        np.testing.assert_array_equal(n_windows_per_sample_, n_windows_per_sample_cum_expected)
